# server/target/openapi.yaml
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Control Finanzas API",
                 "description":  "API para cuentas y ahorro (iteraciÃ³n inicial)",
                 "version":  "0.1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "paths":  {
...
                  "/api/v1/fixed-expenses":  {
                                                 "get":  {
                                                             "tags":  [
                                                                          "fixed-expense-controller"
                                                                      ],
                                                             "operationId":  "list_1",
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "type":  "array",
                                                                                                                                      "items":  {
                                                                                                                                                    "$ref":  "#/components/schemas/FixedExpense"
                                                                                                                                                }
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           }
                                                         },
                                                 "post":  {
                                                              "tags":  [
                                                                           "fixed-expense-controller"
                                                                       ],
                                                              "operationId":  "create_1",
                                                              "requestBody":  {
                                                                                  "content":  {
                                                                                                  "application/json":  {
                                                                                                                           "schema":  {
                                                                                                                                          "$ref":  "#/components/schemas/CreateBody"
                                                                                                                                      }
                                                                                                                       }
                                                                                              },
                                                                                  "required":  true
                                                                              },
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  {
                                                                                                            "*/*":  {
                                                                                                                        "schema":  {
                                                                                                                                       "type":  "object"
                                                                                                                                   }
                                                                                                                    }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          }
                                             },
...
                  "/api/v1/fixed-expenses/{id}":  {
                                                      "delete":  {
                                                                     "tags":  [
                                                                                  "fixed-expense-controller"
                                                                              ],
                                                                     "operationId":  "delete_1",
                                                                     "parameters":  [

                                                                                    ],
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  "OK",
                                                                                                   "content":  {
                                                                                                                   "*/*":  {
                                                                                                                               "schema":  {
                                                                                                                                              "type":  "object"
                                                                                                                                          }
                                                                                                                           }
                                                                                                               }
                                                                                               }
                                                                                   }
                                                                 },
                                                      "patch":  {
                                                                    "tags":  [
                                                                                 "fixed-expense-controller"
                                                                             ],
                                                                    "operationId":  "update",
                                                                    "parameters":  [

                                                                                   ],
                                                                    "requestBody":  {
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/UpdateBody"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    },
                                                                                        "required":  true
                                                                                    },
                                                                    "responses":  {
                                                                                      "200":  {
                                                                                                  "description":  "OK",
                                                                                                  "content":  {
                                                                                                                  "*/*":  {
                                                                                                                              "schema":  {
                                                                                                                                             "type":  "object"
                                                                                                                                         }
                                                                                                                          }
                                                                                                              }
                                                                                              }
                                                                                  }
                                                                }
                                                  },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CreateBody":  {
                                                          "required":  [
                                                                           "accountId",
                                                                           "amount",
                                                                           "categoryId",
                                                                           "date",
                                                                           "type"
                                                                       ],
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "type":  {
                                                                                          "type":  "string",
                                                                                          "enum":  [
                                                                                                       "INCOME",
                                                                                                       "EXPENSE"
                                                                                                   ]
                                                                                      },
                                                                             "accountId":  {
                                                                                               "type":  "integer",
                                                                                               "format":  "int64"
                                                                                           },
                                                                             "categoryId":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            },
                                                                             "amount":  {
                                                                                            "type":  "number",
                                                                                            "format":  "double"
                                                                                        },
                                                                             "date":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                             "description":  {
                                                                                                 "type":  "string"
                                                                                             }
                                                                         }
                                                      },
...
                                       "UpdateBody":  {
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "name":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                             "amount":  {
                                                                                            "type":  "number",
                                                                                            "format":  "double"
                                                                                        },
                                                                             "categoryId":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            }
                                                                         }
                                                      },
...
                                       "Category":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "id":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int64"
                                                                                  },
                                                                           "name":  {
                                                                                        "type":  "string"
                                                                                    }
                                                                       }
                                                    },
...
                                       "FixedExpense":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "id":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int64"
                                                                                      },
                                                                               "name":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                               "amount":  {
                                                                                              "type":  "number"
                                                                                          },
                                                                               "category":  {
                                                                                                "$ref":  "#/components/schemas/Category"
                                                                                            }
                                                                           }
                                                        },
...
                                   }
                   }
}